misc fixes

+++ sys/sys/stddef.h
@@ -62,6 +62,6 @@ typedef	_BSD_WCHAR_T_	wchar_t;
 
 #include <sys/null.h>
 
-#define	offsetof(type, member)	((size_t)(&((type *)0)->member))
+#define	offsetof(type, member)	 ((size_t)(unsigned long)(&(((type *)0)->member)))
 
 #endif /* _SYS_STDDEF_H_ */
